Fix name matching bug in Org Edit form
The edit form in the Organizations module tries to prevent you from
changing an organization's name to be the same as the name of an
existing organization. If the existing organization is the one you are
editing, this should not matter, but the error still appears due to a
bug. This commit fixes that bug.

Signed-off-by: Remington Steed <rjs7@calvin.edu>