fixed the linedFormTable column width
I put an empty row into the linedFormTable to get table-layout:fixed to format with the desired widths. 

This is because the first cell in the first row determines the width when using table-layout:fixed, and the old first row cell spanned two columns and could not have the correct width applied to it without messing up the formatting.