Code cleanup.
DBService is more cleanly organized and is used in conjunction
with DBResult. These two classes are essentially wrappers for
mysql and mysql_result objects. Most of DBService's methods are chainable
as they return a reference to 'this'.

Previously renamed Configuration.php to Config.php and have now removed
the former in git

Reports module no longer dependent on Utility.php, Object.php,
or DatabaseObject.php