Modify the email validation regex
The email address validation was more complex and strict than necessary.
Unless we have a good reason to be more strict, let's use a well tested
pattern. This commit uses the regex from the following web site:

  http://www.regular-expressions.info/email.html

Also, this commit changes the email data encoding function to
encodeURIComponent(), which includes the special characters which are
now allowed in email addresses.

Signed-off-by: Remington Steed <rjs7@calvin.edu>