Fix bug in Org editing existance check
The code provides two ways of excluding the Org being edited when
checking for a duplicate. However, we still see the message "This
organization already exists" when changing the capitalization of an Org
name. This is because both attempts at exclusion are buggy and sometimes
fail. This buggy behavior appears on our older live installation, but
not on our test installation using the latest master code (even though
the buggy code still remains). This commit fixes the bugs anyway (and
removes one of the exclusions to simplify the code), to prevent them
from resurfacing later.

Signed-off-by: Remington Steed <rjs7@calvin.edu>