Remove requirement of Contact Name field
While the code that requires a Contact Name has been present for a long
time, it has done nothing since 2011 when the contactType field was
removed. So a Contact Name hasn't been functionally required since then,
and current consensus is to leave it as not required. This commit
removes the requirement code for that field.

Signed-off-by: Remington Steed <rjs7@calvin.edu>