SQL query building: add quotes to WHERE parameters
To avoid regressions because mysql was fine with: WHERE c.`licenseID`=
whereas mysqli throws a syntax error and needs: WHERE c.`licenseID`=''
This is preventive and not in response to observed regressions.
Also it improves consistency as most of Coral's SQL building use this
form for WHERE parameters.

Part of #54