Incorrect class used when calling getLicenseArray(), breaking right panel in resources view for listing licenses
The list of licenses are not appearing in the right-panel (Helpful Links) in the resources module when viewing a resource.

The problem is that the class method getLicenseArray() was moved out of the Resource class and into the ResourceAcquisition class but the call to this method was not updated in `resources/ajax_htmldata/getRightPanel.php`.

The `resources/resources/cataloging.php` is also using the wrong class object to call getLicenseArray().