(issue #448) Can't switch language if Coral runs on a non-standard port
When running Coral on a non-standard port, for example :9090,
switching the language from English to another is not possible.
When choosing another language from the pull down, the page
reloads and jumps back to English instead of the language chosen.

This is because the lang cookie storing the selected language isn't set.

After some research the problem is that Coral tries to set the
domain attribute of the cookie to the host including the port
number. Instead it should be using the host name without the port.

Cookies do not provide isolation by port. If a cookie is readable
by a service running on one port, the cookie is also readable
by a service running on another port of the same server...
(https://tools.ietf.org/html/rfc6265#section-5.1.3)

Test plan:
- Configure Coral to run on another port, like 9090
- Try to switch to another language than English from different modules and pages
- Check that no lang cookie is stored
- Apply the patch
- Repeat the test plan and verify that the cookie is now
  set properly and switching the language works as expected